axios, api gateway, bug fixes

+++ src/microservices/book/book.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { BookRepo } from './book.repo';
 import { Book } from '../../shared/schemas/book.schema';
 import { EventService } from '../../shared/services/event.service';
+import { APIGateway } from '../../shared/gateway/api_gateway';
 
 @Injectable()
 export class BookService {
@@ -10,7 +11,14 @@ export class BookService {
     private readonly eventService: EventService,
   ) {}
 
-  async upsert(book: Book): Promise<Book | undefined> {
+  async upsert(book: Book, auth: string): Promise<Book | undefined> {
+    const user = await APIGateway.user.authorize(auth);
+
+    if (!user) {
+      console.log(`[I] BookService.upsert(${JSON.stringify(user)}): Unauthorized!`);
+      return undefined
+    }
+    
     const result = await this.repo.upsert(book)
 
     if (result.error) console.log(`[E] BookService.upsert(${JSON.stringify(book)}): ${result.error}`);
@@ -19,7 +27,14 @@ export class BookService {
     return result.data;
   }
 
-  async findById(id:string): Promise<Book | undefined> {
+  async findById(id:string, auth: string): Promise<Book | undefined> {
+    const user = await APIGateway.user.authorize(auth);
+
+    if (!user) {
+      console.log(`[I] BookService.findById(${JSON.stringify(user)}): Unauthorized!`);
+      return undefined
+    }
+
     const result = await this.repo.findById(id);
   
     if (result.error) console.log(`[E] BookService.findById(${JSON.stringify(id)}): ${result.error}`); 
@@ -28,7 +43,14 @@ export class BookService {
     return result.data;
   }
 
-  async find(filter: Record<string, any>): Promise<Book | undefined> {
+  async find(filter: Record<string, any>, auth: string): Promise<Book | undefined> {
+    const user = await APIGateway.user.authorize(auth);
+
+    if (!user) {
+      console.log(`[I] BookService.find(${JSON.stringify(user)}): Unauthorized!`);
+      return undefined
+    }
+
     const result = await this.repo.find(filter);
   
     if (result.error) console.log(`[E] BookService.find(${JSON.stringify(filter)}): ${result.error}`); 
@@ -37,7 +59,14 @@ export class BookService {
     return result.data;
   }
 
-  async findAll(): Promise<Book[] | undefined> {
+  async findAll(auth: string): Promise<Book[] | undefined> {
+    const user = await APIGateway.user.authorize(auth);
+
+    if (!user) {
+      console.log(`[I] BookService.findAll(${JSON.stringify(user)}): Unauthorized!`);
+      return undefined
+    }
+
     const result = await this.repo.findAll();
 
     if (result.error) console.log(`[E] BookService.findAll(): ${result.error}`); 
@@ -46,7 +75,14 @@ export class BookService {
     return result.data;
   }
 
-  async delete(id: string): Promise<Book | undefined> {
+  async delete(id: string, auth: string): Promise<Book | undefined> {
+    const user = await APIGateway.user.authorize(auth);
+
+    if (!user) {
+      console.log(`[I] BookService.delete(${JSON.stringify(user)}): Unauthorized!`);
+      return undefined
+    }
+
     const result = await this.repo.delete(id);
 
     if (result.error) console.log(`[E] BookService.delete(${id}): ${result.error}`);
