+++ src/microservices/rating/rating.service.ts
@@ -3,26 +3,32 @@ import { RatingRepo } from './rating.repo';
 import { EventService } from '../../shared/services/event.service';
 import { Rating } from '../../shared/schemas/rating.schema';
 import { UpsertRatingDTO } from './dto/upsert-rating.dto';
-import { BookService } from '../book/book.service';
 import { User } from '../../shared/schemas/user.schema';
 import { Types } from 'mongoose';
+import { APIGateway } from '../../shared/gateway/api_gateway';
 
 @Injectable()
 export class RatingService {
   constructor(
     private readonly repo: RatingRepo,
-    private readonly bookService: BookService,
     private readonly eventService: EventService,
   ) {}
 
-  async upsert(dto: UpsertRatingDTO, user: User): Promise<Rating | undefined> {
-    const book = await this.bookService.find({ _id: dto.bookId })
+  async upsert(dto: UpsertRatingDTO, auth: string): Promise<Rating | undefined> {
+    const user = await APIGateway.user.authorize(auth);
+
+    if (!user) {
+      console.log(`[I] RatingService.upsert(${JSON.stringify(user)}): Unauthorized!`);
+      return undefined
+    }
+
+    const book = await APIGateway.book.getBookById(auth, dto.bookId)
     
     if (book) {
       const result = await this.repo.upsert({
         _id: dto._id ? new Types.ObjectId(dto._id) : undefined,
-        book,
-        user,
+        book: new Types.ObjectId(book._id),
+        user: new Types.ObjectId(user._id),
         rating: dto.rating,
         review: dto.review,
       } as Rating);
@@ -36,7 +42,14 @@ export class RatingService {
     }
   }
 
-  async find(filter: Record<string, any>): Promise<Rating | undefined> {
+  async find(filter: Record<string, any>, auth: string): Promise<Rating[] | undefined> {
+    const user = await APIGateway.user.authorize(auth);
+
+    if (!user) {
+      console.log(`[I] RatingService.find(${JSON.stringify(user)}): Unauthorized!`);
+      return undefined
+    }
+
     const result = await this.repo.find(filter);
   
     if (result.error) console.log(`[E] RatingService.find(${JSON.stringify(filter)}): ${result.error}`); 
@@ -45,16 +58,14 @@ export class RatingService {
     return result.data;
   }
 
-  async findAll(): Promise<Rating[] | undefined> {
-    const result = await this.repo.findAll();
+  async delete(id: string, auth: string): Promise<Rating | undefined> {
+    const user = await APIGateway.user.authorize(auth);
 
-    if (result.error) console.log(`[E] RatingService.findAll(): ${result.error}`); 
-    else console.log(`[I] RatingService.findAll(): ${JSON.stringify(result.data)}`);
+    if (!user) {
+      console.log(`[I] RatingService.delete(${JSON.stringify(user)}): Unauthorized!`);
+      return undefined
+    }
 
-    return result.data;
-  }
-
-  async delete(id: string): Promise<Rating | undefined> {
     const result = await this.repo.delete(id);
 
     if (result.error) console.log(`[E] RatingService.delete(${id}): ${result.error}`);
